Third review pass: remove what the second pass left
ADDR was a cast wrapper whose three uses each cast the result to something else, so the cast
it performed was never the one wanted.

mesh-stat wrote the four page-state names as eight string literals beside the enum that
already names them, so changing NOWN would have broken the JSON silently; the names are a
table and the fields print in a loop. Its two failure paths printed the same line.

The client clamped a length to the payload longhand while mesh_clamp existed, the same fact
at two types; mesh_clamp now takes size_t and both sides use it. mesh_pay was recomputed per
iteration of the write loop and twice in open.

A completion whose work-request id fell outside the pool was skipped in silence. The
contract calls a branch that quietly does less worse than one that fails loudly, and this one
cannot happen unless the identifier we set has been corrupted, so it dies.

+++ rdma/mesh.h
@@ -30,8 +30,8 @@ static inline unsigned char *mesh_data(struct hdr *m, uint32_t i){
   return mesh_at(m,i) + sizeof(struct wire); }
 static inline uint32_t mesh_pay(struct hdr *m){
   return m->pgsz - (uint32_t)sizeof(struct wire); }
-static inline uint32_t mesh_clamp(struct hdr *m, uint32_t n){
-  uint32_t p=mesh_pay(m); return n>p?p:n; }
+static inline uint32_t mesh_clamp(struct hdr *m, size_t n){
+  uint32_t p=mesh_pay(m); return n>p?p:(uint32_t)n; }
 static inline struct desc *slot(struct hdr *m, int k, uint64_t i){
   return &((struct desc*)((unsigned char*)m + RINGS))[k*MESH_RING + i%MESH_RING]; }
 static inline int push(struct hdr *m, int k, const struct desc *d){
